requirements.txt
mido==1.2.10
python-rtmidi==1.5.8
setuptools>=65.5.0

